requirements.txt
ultralytics==8.0.132
matplotlib==3.7.1
numpy==1.22.4
opencv-python==4.7.0.72
torch==2.0.1
torchvision==0.15.2
argparse==1.4.0
Pillow==9.3.0